+++ test/rails_app/app/controllers/application_controller.rb
@@ -1,6 +1,8 @@
 class ApplicationController < ActionController::Base
   protect_from_forgery with: :exception
 
+  private
+
   def current_account
     Account.find(rodauth.session_value)
   end